corrected some logging, fixed the incorrect flagging matterial. I think... I won't know for sure until the live test tomorrow.
main changes were in tll_trading_strategy.py, the changes weren't even in teh identify_flags() function, where the changes were was in the check for bearish and bullish breakout functions. I didn't know when coding those functions that I only had it detect lower highs (for bullish, LH) and higher_lows (for bearish, HL) and it would only check the closing price if the candle was a LH or HL. so i added a if statement outside of the LH/HL section that checks the closing price. luckily that worked. then i ran into misssing slope and intercept errors, so to handle that i added another if statement near the tops of bearish and bullish breakout check function. if slope and intercept are not none, then calculate trendline_y or the slope intercept point AKA Y=Mx+B.

This only took me a day to figure out, the way i figured out this problem was through deductive reasoning. I places a bunched on print statement everywhere. and then when the occurance happened I look to see which print statements were triggered which got me closer and closer to the needle in the haystack.