there was a incorrect buy that happened today and hopefully i fixed it.
it kept on checking if it was above emas and tried to see if it was above a line that was not complete, it had point_2 but not point_1 and it made the line_data status complete when it truely was not. Hopefully this fixed that.

The only Code change i made was in the tll_trading_strategy.py

 the bot has made 2 trades currently, we will see at end of day. first trade was incorrect, but 2nd trade was. Would be up 64$ if strategy worked as it should.