tll, just giving it a description, it doesn't use it for anything, atleast not yet. also added a whole pass along thing for 'what_type_of_candle' so we can see where our orders are coming from.
for data_aquisition.py, thats where most of todays work was done. the important stuff is at the bottom of the script. where we have 2 function that add and read duplicates in a json file that i created called 'order_candle_type.json' that is meant to keep track of how many orders happen in a given zone, if an order takes place and we see that we have already made an order in that given zone mutliple times then it cancels the buy because most traders start to see the pattern happen and they get in and thats where things get all fucky-waky.

main.py made it reset that json file

test.py, i moved a function/variable in a different script so i had to change import settings.