added more resetting of values to the tll trading strat because i think when orders are active or about to be, the values get carried over when they dont need to be, confuses the program and it seemed to have worked somewhat today.
i still have some flag issues with plotting them, i wrote some new code in chart_visualizations to help visualize it so i can see the problem. and i have another computer thats going to run a different version of code to see if that works and if it does then i transfer it over.