changing account size back to yestardays numbers since todays order was a incorrect order.
chart_visualization update_2_min() needed some changes.

tll_trading_strategy, data_aquisition: i added the update_2_min to get a better understanding of what is going on and why the flagging is working incorrectly.

Actually i know why its not working but i know that i need to add another check to the balances but i don't know what i should add without conflicting with other parts of the program. this will take a while to think of a solution