HUGE UPDATE:
I created a new simulator to test flag code and i have something really good down.

alot of code was added to tll, chart_vis, and data_aqua, main and so on. flags are much more reliable now. tomorrow is going to be a fun run.

The ammount of data i used to test these flags is insane. for the paste month or something like that i had discord save the end of day results incase i need to test some new processes with it and today was that day.