week 11 day 1 report:
Beginning of day had alot of errors that i had to fix. dealing with the new boxes being plotted correctly and sending data correctly to other scripts. I did have a error on candle_zone_handler() in tll_trading_strategy.py, no errors came up it just wasn't working correctly. Plus order handling had issues when 13 ema broke.

There was a total of 4 wrong trades but I did restart the script in the middle of the day, thats why the script didn't record all of the trades. the only trade that was valid was the second trade.