week 11 day 2 report:
I had to change candle_zone_handler() to see if boxes even exist because it would raise a error if they didn't exist. Also had to change zone type because i realise that most of the code i wrote before hand of all this updated code was meant for one PDHL when now with all this new boxes i have multiple PDHL's and there all named differently as they should. just fixing it all to handle the new updates.