week 12 day 4 report
I removed the 3 candles into zone to see if it would get out of zone but cross a flag, its caused more errors than what its worth.

Plus im starting to realise that its not part of the strat, the strat is that price comes out of the zone, flag forms and then breaks out and we buy.