Week 16 day 4 report:
Had Alot of errors today,

first error was with 'boxes.py': the PDHL's, some of them were not being numbered correctly where the top position of the box was the bottom and vise-versa with the bottom. found the spots and made sure the top/bottom values were put in the right spot. line 350-352 were the main culperate.

data_aquisition.py: had alot of internet issues today, that was the main culperate of vast majority of the errors today. added line 100-104 for more redundency and less chatter in the terminal/logs.

tll_trading_strategy.py: line 187 had to be commented out, it was messing with the enter/exit calculations with the price action of the zones. it would stop it from calculating.

order_handler.py: because of the internet being very slow today, had to add some contigincies if i was in a order because i aws today and it messed up the discord order messages (fixed them with the edit_discord_message.py). main fix was line 612-614 because i don't want it to have a retry limit, i want it to keep on trying until the internet comes back on. second main fix was line 143-178, it would get bid price and if it failed it shat itself, where now it keeps on getting bid price until it has it then returns it. I also cleaned alot of the get_message_content() so it looks better.

edit_discord_message.py: its to help me change the message on discord because today i had some internet issue which let to the orders not logging correctly, so i had to fix those messages so the end of day calculations doesn't mess up. ended up having to edit that message too.

This is alot of code that i changed and i hope it work cohercivly tomorrow. if it doesn't its back to debugging.