Week 29 day 1 report
tll_trading_strategy.py: had some issues, nothing crazy, when it switchted to after 15 minutes it duplicated ema data, i think i fixed it but wont know until tomorrow. having some flag identification issues, it looks like the cause of it was the ladder effeect of passing 'what_type_of_candle' through a bunch of functions and forgetting what does what and it turns out the begining process was right but then the reset on the start_new_flag_values() set the zone type to the wrong value. I also forgot to change the config file which led to a flag angle error. Forgot to change some of the values being imported into process_breakout_detection(). forgot to change some of the start_new_flag_values() to what i set previously, fixed that.

Had to run/restart a bunch of times.